Added stepPrecision that defaults to step * 0.1. stepPrecision is used when holding shift
Breaking change where we've inverted the behaviour (that's not commonly used) for holding shift.

The approach we are going for now is that step should be a 'reasonable' to change the number value by default given its use case and holding shift will give precision.

The stepPrecision is defaulted to step * 0.1 but can be customised on a per input basis to be something else if we choose to